--- patch/models.py
@@ -1,10 +1,0 @@
-# Create your models here.
-from django.db import models
-from django.utils import timezone
-
-class PatchPost(models.Model):
-    title = models.CharField(max_length=256)
-    content = models.TextField()
-    created_at = models.DateTimeField(default=timezone.now)
-    def __str__(self):
-        return self.title
