--- patch/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework.serializers import ModelSerializer
-from .models import PatchPost
-
-class PatchSerializer(ModelSerializer):
-    class Meta:
-        model = PatchPost
-        fields = ['title', 'content']
