+++ post/serializers.py
@@ -5,3 +5,4 @@ class PostSerializer(ModelSerializer):
     class Meta:
         model = Post
         fields = "__all__"
+
