+++ post/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 ### 추가
-from .views import PostListView, PostDetailView
+from .views import PostListView, PostDetailView, PostUpdateView
 ###
 
 app_name = 'post'
@@ -10,7 +10,7 @@ app_name = 'post'
 # 예) api/post/3이라면 <int:post_id>로
 urlpatterns = [
     # CBV url path
-    path("", PostListView.as_view()), ### 추가
-    path("<int:post_id>/", PostDetailView.as_view()), ### 추가
-
+    path("", PostListView.as_view()),
+    path("<int:post_id>/", PostDetailView.as_view()),
+    path("<int:post_id>/update", PostUpdateView()),
 ]
