+++ seminar/settings.py
@@ -49,7 +49,6 @@ INSTALLED_APPS = [
     'post',
     'django.contrib.staticfiles',  # required for serving swagger ui's css/js files
     'drf_yasg',
-    'patch',
 ]
 
 MIDDLEWARE = [
