+++ seminar/urls.py
@@ -19,7 +19,6 @@ schema_view = get_schema_view(
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/post/', include('post.urls')),
-    path('api/patch/', include('patch.urls')),
     # swagger path
     path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
 ]
